Adjust relay code to include message in notifications to prevent overuse of virtual, store message type inside packets to retrieve the message from stored relay packet

+++ src/TAO/API/types/users/logout.cpp
@@ -81,7 +81,7 @@ namespace TAO
             {
                 /* Check whether it is valid before relaying it to all peers */
                 if(ssMessage.size() > 0)
-                    LLP::TRITIUM_SERVER->Relay(LLP::ACTION::DEAUTH, ssMessage.Bytes());
+                    LLP::TRITIUM_SERVER->Relay(uint8_t(LLP::ACTION::DEAUTH), ssMessage.Bytes());
 
                 /* Free up the Auth private key */
                 pAuthKey.free();
